bot/src/requirements.txt
requests
botocore
python-telegram-bot